+++ test/git/pkgs/test_ecosystems_client.rb
@@ -116,4 +116,73 @@ class Git::Pkgs::TestEcosystemsClient < Minitest::Test
       @client.bulk_lookup(["pkg:gem/rails"])
     end
   end
+
+  def test_lookup_all_versions
+    stub_request(:get, "https://packages.ecosyste.ms/api/v1/registries/rubygems.org/packages/rails/versions?page=1&per_page=100")
+      .to_return(
+        status: 200,
+        body: [
+          { "number" => "7.0.0", "published_at" => "2021-12-15T00:00:00Z" },
+          { "number" => "7.1.0", "published_at" => "2023-10-05T00:00:00Z" }
+        ].to_json,
+        headers: { "Content-Type" => "application/json" }
+      )
+
+    versions = @client.lookup_all_versions("pkg:gem/rails")
+
+    assert_equal 2, versions.size
+    assert_equal "7.0.0", versions[0]["number"]
+    assert_equal "7.1.0", versions[1]["number"]
+  end
+
+  def test_lookup_all_versions_paginates
+    stub_request(:get, "https://packages.ecosyste.ms/api/v1/registries/rubygems.org/packages/rails/versions?page=1&per_page=100")
+      .to_return(
+        status: 200,
+        body: (1..100).map { |i| { "number" => "1.0.#{i}" } }.to_json,
+        headers: { "Content-Type" => "application/json" }
+      )
+
+    stub_request(:get, "https://packages.ecosyste.ms/api/v1/registries/rubygems.org/packages/rails/versions?page=2&per_page=100")
+      .to_return(
+        status: 200,
+        body: [{ "number" => "2.0.0" }].to_json,
+        headers: { "Content-Type" => "application/json" }
+      )
+
+    versions = @client.lookup_all_versions("pkg:gem/rails")
+
+    assert_equal 101, versions.size
+  end
+
+  def test_lookup_all_versions_returns_nil_for_not_found
+    stub_request(:get, %r{packages.ecosyste.ms/api/v1/registries/rubygems.org/packages/nonexistent/versions})
+      .to_return(status: 404)
+
+    result = @client.lookup_all_versions("pkg:gem/nonexistent")
+
+    assert_nil result
+  end
+
+  def test_bulk_lookup_all_versions
+    stub_request(:get, %r{packages.ecosyste.ms/api/v1/registries/rubygems.org/packages/rails/versions})
+      .to_return(
+        status: 200,
+        body: [{ "number" => "7.1.0", "published_at" => "2023-10-05T00:00:00Z" }].to_json,
+        headers: { "Content-Type" => "application/json" }
+      )
+
+    stub_request(:get, %r{packages.ecosyste.ms/api/v1/registries/npmjs.org/packages/lodash/versions})
+      .to_return(
+        status: 200,
+        body: [{ "number" => "4.17.21", "published_at" => "2021-02-20T00:00:00Z" }].to_json,
+        headers: { "Content-Type" => "application/json" }
+      )
+
+    results = @client.bulk_lookup_all_versions(["pkg:gem/rails", "pkg:npm/lodash"])
+
+    assert_equal 2, results.size
+    assert_equal 1, results["pkg:gem/rails"].size
+    assert_equal 1, results["pkg:npm/lodash"].size
+  end
 end